Address reviewer comments: move MG code from MooseApp
refs #23352
