+++ framework/include/base/MooseApp.h
@@ -22,6 +22,7 @@
 #include "PerfGraphInterface.h"
 #include "TheWarehouse.h"
 #include "RankMap.h"
+#include "MeshGeneratorSystem.h"
 
 #include "libmesh/parallel_object.h"
 #include "libmesh/mesh_base.h"
@@ -40,7 +41,6 @@ class Executor;
 class NullExecutor;
 class Backup;
 class FEProblemBase;
-class MeshGenerator;
 class InputParameterWarehouse;
 class SystemInfo;
 class CommandLine;
@@ -723,26 +723,9 @@ public:
   const MooseMesh * masterDisplacedMesh() const { return _master_displaced_mesh; }
 
   /**
-   * Class that is used as a parameter to setFinalMeshGeneratorName()
-   * that allows only MeshGeneratorMesh methods to call this methods
+   * Gets the system that manages the MeshGenerators
    */
-  class SetFinalMeshGeneratorNameKey
-  {
-    friend class MeshGeneratorMesh;
-    SetFinalMeshGeneratorNameKey() {}
-    SetFinalMeshGeneratorNameKey(const SetFinalMeshGeneratorNameKey &) {}
-  };
-
-  /**
-   * Set final mesh generator name
-   *
-   * Guarded by the SetFinalMeshGeneartorNameKey so that it can only be called
-   * by MeshGeneratorMesh. This works because the copy construction of the
-   * key class happens within the caller, where the copy constructor
-   * is private and access given via a friend.
-   */
-  void setFinalMeshGeneratorName(const std::string & generator_name,
-                                 const SetFinalMeshGeneratorNameKey);
+  MeshGeneratorSystem & getMeshGeneratorSystem() { return _mesh_generator_system; }
 
   /**
    * Add a mesh generator that will act on the meshes in the system
@@ -751,52 +734,40 @@ public:
    * @param name The name of the MeshGenerator
    * @param params The params used to construct the MeshGenerator
    *
-   * Internally, this will store the parameters for future construction
-   * during the "add_mesh_generator" task. When called during the
-   * "create_mesh_generator" task (i.e., when creating mesh subgenerators),
-   * it will also construct the generator.
-   *
-   * We don't construct them yet because we want to create them in order
-   * during createMeshGenerators() as much as possible so that we don't
-   * need lazy construction for things like mesh properties.
+   * See MeshGeneratorSystem::addMeshGenerator()
    */
   void addMeshGenerator(const std::string & type,
                         const std::string & name,
-                        const InputParameters & params);
-  /**
-   * Get a mesh generator with its name
-   */
-  const MeshGenerator & getMeshGenerator(const std::string & name) const;
+                        const InputParameters & params)
+  {
+    _mesh_generator_system.addMeshGenerator(type, name, params);
+  }
 
   /**
-   * Get names of all mesh generators
-   * Note: This function should be called after all mesh generators are added with the
-   * 'add_mesh_generator' task. The returned value will be undefined and depends on the ordering
-   * that mesh generators are added by MOOSE if the function is called during the
-   * 'add_mesh_generator' task.
+   * @returns Whether or not a mesh generator exists with the name \p name.
    */
-  std::vector<std::string> getMeshGeneratorNames() const;
+  bool hasMeshGenerator(const MeshGeneratorName & name) const
+  {
+    return _mesh_generator_system.hasMeshGenerator(name);
+  }
 
   /**
-   * Whether or not a mesh generator exists with the given name
+   * @returns The MeshGenerator with the name \p name.
    */
-  bool hasMeshGenerator(const MeshGeneratorName & name) const;
+  const MeshGenerator & getMeshGenerator(const std::string & name) const
+  {
+    return _mesh_generator_system.getMeshGenerator(name);
+  }
 
   /**
-   * Whether or not we know about the parameters for a MeshGenerator with the given name
+   * @returns The names of all mesh generators
    *
-   * This is primarily for error checking. If MeshGenerator dependencies are screwed up,
-   * someone could be looking for a MeshGenerator that hasn't been constructed yet.
-   * With this, at least we can give the user some context that we know the generator
-   * exists, just that the dependencies are hosed.
+   * See MeshGeneratorSystem::getMeshGeneratorNames()
    */
-  bool hasMeshGeneratorParams(const MeshGeneratorName & name) const;
-
-  /**
-   * Get a refernce to a pointer that will be the output of the
-   * MeshGenerator named name
-   */
-  [[nodiscard]] std::unique_ptr<MeshBase> & getMeshGeneratorOutput(const MeshGeneratorName & name);
+  std::vector<std::string> getMeshGeneratorNames() const
+  {
+    return _mesh_generator_system.getMeshGeneratorNames();
+  }
 
   /**
    * Append a mesh generator that will act on the final mesh generator in the system
@@ -805,27 +776,13 @@ public:
    * @param name The name of the MeshGenerator
    * @param params The params used to construct the MeshGenerator
    *
-   * This MeshGenerator must have a parameter "input" of type MeshGeneratorName
-   * for this to work, as said parameter is set to the current final generator
-   *
-   * Note: This function must be called during the append_mesh_generator task.
+   * See MeshGeneratorSystem::appendMeshGenerator()
    */
   const MeshGenerator &
-  appendMeshGenerator(const std::string & type, const std::string & name, InputParameters params);
-
-  /**
-   * Execute and clear the Mesh Generators data structure
-   */
-  void executeMeshGenerators();
-
-  /**
-   * Creates (constructs) all of the MeshGenerators that been
-   * declared using addMeshGenerator().
-   *
-   * Should only be called by the CreateAddedMeshGenerators during
-   * the create_added_mesh_generators task.
-   */
-  void createAddedMeshGenerators();
+  appendMeshGenerator(const std::string & type, const std::string & name, InputParameters params)
+  {
+    return _mesh_generator_system.appendMeshGenerator(type, name, params);
+  }
 
   /**
    * Whether this app is constructing mesh generators
@@ -834,16 +791,6 @@ public:
    * construct MeshGenerators in unit tests
    */
   virtual bool constructingMeshGenerators() const;
-
-  /**
-   * Whether or not this app is appending mesh generators (append_mesh_generator task)
-   */
-  bool appendingMeshGenerators() const;
-
-  /**
-   * Get the generated mesh generated by executeMeshGenerators();
-   */
-  std::unique_ptr<MeshBase> getMeshGeneratorMesh(bool check_unique = true);
 
   ///@{
   /**
@@ -1264,11 +1211,6 @@ private:
   void setCheckUnusedFlag(bool warn_is_error = false);
 
   /**
-   * Order all of the _mesh_generators into _ordered_mesh_generators
-   */
-  void createMeshGeneratorOrder();
-
-  /**
    * @return whether we have created any clones for the provided template relationship manager and
    * mesh yet. This may be false for instance when we are in the initial add relationship manager
    * stage and haven't attempted attaching any relationship managers to the mesh or dof map yet
@@ -1339,33 +1281,6 @@ private:
    */
   bool runInputs() const;
 
-  /**
-   * Get a MeshGenerator with the name \p name.
-   *
-   * We add the "internal" here so that we allow objects that have non-const access to
-   * MooseApp to call getMeshGenerator without a const_cast. If the name was the same,
-   * you'd get an error about trying to access a private method.
-   */
-  MeshGenerator & getMeshGeneratorInternal(const std::string & name)
-  {
-    return const_cast<MeshGenerator &>(std::as_const(*this).getMeshGenerator(name));
-  }
-
-  /**
-   * Gets the MeshGeneratorNames that are referenced in an object's parameters.
-   *
-   * The result is returned as a pair of param name -> MeshGeneratorName in order
-   * to provide context.
-   */
-  std::vector<std::pair<std::string, MeshGeneratorName>>
-  getMeshGeneratorParamDependencies(const InputParameters & params) const;
-
-  /**
-   * Internal method for actually constructing a mesh generator after it
-   * has been delcared externally in addMeshGenerator (in Actions).
-   */
-  std::shared_ptr<MeshGenerator> createMeshGenerator(const std::string & name);
-
   /// General storage for custom RestartableData that can be added to from outside applications
   std::unordered_map<RestartableDataMapName, std::pair<RestartableDataMap, std::string>>
       _restartable_meta_data;
@@ -1392,24 +1307,8 @@ private:
   /// The displaced mesh from master app
   const MooseMesh * const _master_displaced_mesh;
 
-  /// The MeshGenerators declared using addMeshGenerator(), cleared after createMeshGenerators()
-  /// Key is the name, pair contains the type and the params
-  std::unordered_map<std::string, std::pair<std::string, InputParameters>> _mesh_generator_params;
-
-  /// Holds the mesh generators until they are executed, then this structure is cleared
-  std::map<std::string, std::shared_ptr<MeshGenerator>> _mesh_generators;
-
-  /// Holds the ordered mesh generators from createMeshGeneratorOrder() until they are executed
-  std::vector<std::vector<MeshGenerator *>> _ordered_mesh_generators;
-
-  /// Holds the output for each mesh generator - including duplicates needed downstream
-  std::map<std::string, std::list<std::unique_ptr<MeshBase>>> _mesh_generator_outputs;
-
-  /// The final mesh generator name to use
-  std::string _final_generator_name;
-
-  /// The final Mesh that is generated by the generators
-  std::list<std::unique_ptr<MeshBase> *> _final_generated_meshes;
+  /// The system that manages the MeshGenerators
+  MeshGeneratorSystem _mesh_generator_system;
 
   /// Cache for a Backup to use for restart / recovery
   std::shared_ptr<Backup> _cached_backup;
@@ -1419,10 +1318,6 @@ private:
 
   /// Whether to turn on automatic scaling by default
   const bool _automatic_automatic_scaling;
-
-  /// Whether the mesh generator MeshBase has been popped off its storage container and is no
-  /// longer accessible
-  bool _popped_final_mesh_generator;
 
   /// CPU profiling
   bool _cpu_profiling = false;
