+++ framework/src/mesh/MeshGeneratorMesh.C
@@ -29,8 +29,9 @@ MeshGeneratorMesh::validParams()
 MeshGeneratorMesh::MeshGeneratorMesh(const InputParameters & parameters) : MooseMesh(parameters)
 {
   if (isParamValid("final_generator"))
-    _app.setFinalMeshGeneratorName(getParam<std::string>("final_generator"),
-                                   MooseApp::SetFinalMeshGeneratorNameKey());
+    _app.getMeshGeneratorSystem().setFinalMeshGeneratorName(
+        getParam<std::string>("final_generator"),
+        MeshGeneratorSystem::SetFinalMeshGeneratorNameKey());
 }
 
 std::unique_ptr<MooseMesh>
@@ -43,5 +44,5 @@ void
 MeshGeneratorMesh::buildMesh()
 {
   if (!hasMeshBase())
-    _mesh = _app.getMeshGeneratorMesh();
+    _mesh = _app.getMeshGeneratorSystem().getMeshGeneratorMesh();
 }
